[implement]: Post CRUD 구현 - get /api/posts/{post-name} - get /api/posts/ - get /api/user/posts - post /api/user/posts - put /api/user/posts - delete /api/user/posts

+++ src/main/java/com/yemin/twitter/repository/MemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<MemberVO, Integer> {
+public interface MemberRepository extends JpaRepository<MemberVO, Long> {
 
     public boolean existsByEmail(String email);
